Add loading indicators to admin photo menu
